Fix BrTextStylerSyncStrategy
I had an error because #text was sent from another thread.
Use #textSnapshot instead of #text to ensure it's executed from
the right text. This is the "sync" part of this implementation.

But for dispatching the BrTextStyledEvent, I'm changing to
do it from an enqueued task. Not sure this is needed. I think
this doesn't break the "sync" intention of the class.